+++ CPUTuneApp/CPUTuneApp/ContentView.swift
@@ -1,0 +1,23 @@
+//
+//  ContentView.swift
+//  CPUTuneApp
+//
+//  Created by Yating Zhou on 7/5/20.
+//  Copyright © 2020 syscl. All rights reserved.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        Text("Hello, World!")
+            .frame(maxWidth: .infinity, maxHeight: .infinity)
+    }
+}
+
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}