+++ CPUTuneCore/CPUTune/CPUTune.cpp
@@ -24,13 +24,14 @@ IOService *CPUTune::probe(IOService *provider, SInt32 *score) {
 
 bool CPUTune::init(OSDictionary *dict)
 {
+    LOG("CPUTune(%s) starting on macOS Darwin %d.%d.", kmod_info.version, getKernelVersion(), getKernelMinorVersion());
     if (getKernelVersion() >= KernelVersion::Unsupported && !checkKernelArgument(bootargBeta)) {
-        myLOG("Unsupported kernel version: %d, get a CPUTune that supports current kernel from https://github.com/syscl/CPUTune", getKernelVersion());
+        LOG("Unsupported kernel version: %d, get a CPUTune that supports current kernel from https://github.com/syscl/CPUTune", getKernelVersion());
         nvram.setKextPanicKey();
         return false;
     } else if (nvram.isKextPanicLastBoot()) {
         // clear the panic key
-        myLOG("Found % key being set in NVRAM, CPUTune (%s) supportted kernel version %d, clear the panic key", kCPUTUNE_PANIC_KEY, kextVersion, getKernelVersion());
+        LOG("Found %s key being set in NVRAM, CPUTune(%s) supportted kernel version %d, clear the panic key", kCPUTUNE_PANIC_KEY, kmod_info.version, getKernelVersion());
         nvram.clearKextPanicKey();
     }
     
@@ -38,13 +39,12 @@ bool CPUTune::init(OSDictionary *dict)
                       checkKernelArgument("-x") |
                       checkKernelArgument(bootargOff);
     if (isDisabled) {
-        myLOG("init: not allow to run.");
+        LOG("not allow to run.");
         return false;
     }
-    bool ret = super::init(dict);
-    if (!ret) {
-        myLOG("init: failed!");
-        return ret;
+    if (!super::init(dict)) {
+        LOG("super init failed!");
+        return false;
     }
     
     // get string properties
@@ -61,18 +61,20 @@ bool CPUTune::init(OSDictionary *dict)
     
     if (OSNumber *timeout = OSDynamicCast(OSNumber, getProperty("UpdateInterval"))) {
         updateInterval = timeout->unsigned32BitValue();
-        myLOG("Update time interval %u ms per cycle", updateInterval);
+        LOG("Update time interval %u ms per cycle", updateInterval);
     }
-    
-    myLOG("init: succeeded!");
     
     org_MSR_IA32_MISC_ENABLE = rdmsr64(MSR_IA32_MISC_ENABLE);
     org_MSR_IA32_PERF_CTL = rdmsr64(MSR_IA32_PERF_CTL);
     org_MSR_IA32_POWER_CTL = rdmsr64(MSR_IA32_POWER_CTL);
-    org_HWPRequest = rdmsr64(MSR_IA32_HWP_REQUEST);
+    if (cpu_info.supportedHWP) {
+        // Do not read MSR_IA32_HWP_REQUEST on unsupported HWP's CPU (otherwise will cause panic)
+        org_HWPRequest = rdmsr64(MSR_IA32_HWP_REQUEST);
+    }
     org_TurboRatioLimit = rdmsr64(MSR_TURBO_RATIO_LIMIT);
     
-    return ret;
+    LOG("succeeded!");
+    return true;
 }
 
 const char* CPUTune::getStringPropertyOrElse(const char* key, const char* defaultProperty) const {
@@ -91,10 +93,9 @@ const bool CPUTune::getBooleanOrElse(const char* key, const bool defaultValue) c
 
 bool CPUTune::start(IOService *provider)
 {
-    bool ret = super::start(provider);
-    if (!ret || provider == nullptr) {
-        myLOG("start: cannot start provider or provider does not exist.");
-        return ret;
+    if (!super::start(provider) || provider == nullptr) {
+        LOG("cannot start provider or provider does not exist.");
+        return false;
     }
     
     // let's turn off some of the SIP bits so that we can debug it easily on a real mac
@@ -107,18 +108,18 @@ bool CPUTune::start(IOService *provider)
     timerSource = IOTimerEventSource::timerEventSource(this,
                     OSMemberFunctionCast(IOTimerEventSource::Action, this, &CPUTune::readConfigAtRuntime));
     if (!timerSource) {
-        myLOG("start: failed to create timer event source!");
+        LOG("failed to create timer event source!");
         // Handle error (typically by returning a failure result).
         return false;
     }
     
     if (myWorkLoop->addEventSource(timerSource) != kIOReturnSuccess) {
-        myLOG("start: failed to add timer event source to work loop!");
+        LOG("failed to add timer event source to work loop!");
         // Handle error (typically by returning a failure result).
         return false;
     }
     
-    timerSource->setTimeoutMS(2000);
+    timerSource->setTimeoutMS(updateInterval);
     
     // check if we need to enable Intel Turbo Boost
     if (enableIntelTurboBoost) {
@@ -130,10 +131,8 @@ bool CPUTune::start(IOService *provider)
     // make sure we disable ProcHot only if turboboost is disabled
     if (enableIntelProcHot) {
         enableProcHot();
-    } else if(!enableIntelTurboBoost) {
-        disableProcHot();
     } else {
-        myLOG("start: cannot deactivate PROCHOT while turboboost is active!");
+        disableProcHot();
     }
     
     // check if we need to enable Intel Speed Shift on platform on Skylake+
@@ -146,12 +145,12 @@ bool CPUTune::start(IOService *provider)
             hwpEnableOnceSet = true;
         } 
     } else {
-        myLOG("start: cpu model (0x%x) does not support Intel SpeedShift.", cpu_info.model);
+        LOG("cpu model (0x%x) does not support Intel SpeedShift.", cpu_info.model);
     }
     
-    myLOG("start: registerService");
+    LOG("registerService");
     registerService();
-    return ret;
+    return true;
 }
 
 void CPUTune::readConfigAtRuntime(OSObject *owner, IOTimerEventSource *sender)
@@ -164,99 +163,80 @@ void CPUTune::readConfigAtRuntime(OSObject *owner, IOTimerEventSource *sender)
     // As for now, the reading procedure reads only one byte, which is fairly fast in our case, so we assume
     // this routine will not cause infinite blocking. Let me know if you have some other good ideas.
     if (turboBoostPath) {
-        // check if previous turbo boost is enabled
-        bool prev = rdmsr64(MSR_IA32_MISC_ENABLE) == (org_MSR_IA32_MISC_ENABLE & kEnableTurboBoostBits);
-        uint8_t *buffer = readFileNBytes(turboBoostPath, 0, 1);
-        // check if currently request enable turbo boost
-        bool curr = buffer && (*buffer == '1');
-        // deallocate the buffer
-        deleter(buffer);
-        if (curr != prev) {
-            myLOG("readConfigAtRuntime: %s Intel Turbo Boost", curr ? "enable" : "disable");
-            if (curr) {
+        if (uint8_t *buffer = readFileAsBytes(turboBoostPath, 0, 1)) {
+            if (*buffer == '1') {
                 enableTurboBoost();
             } else {
                 disableTurboBoost();
-
             }
+            deleter(buffer);
         }
     }
     
     // Turbo ratio limit
-    if ((rdmsr64(MSR_IA32_MISC_ENABLE) & kEnableTurboBoostBits) && turboRatioLimitConfigPath) {
+    if ((rdmsr64(MSR_IA32_MISC_ENABLE) & kEnableTurboBoostBits) && cpu_info.turboRatioLimitRW && turboRatioLimitConfigPath) {
         size_t valid_length = cpu_info.coreCount * 2 + 2; // +2 for '0x'/'0X'
-        if (uint8_t* config = readFileNBytes(turboRatioLimitConfigPath, 0, valid_length)) {
+        if (uint8_t* config = readFileAsBytes(turboRatioLimitConfigPath, 0, valid_length)) {
             long limit = hexToInt(reinterpret_cast<char*>(config));
             deleter(config);
             if (limit == ERANGE) {
-                myLOG("%s: Turbo ratio limit is not a valid hexadecimal constant at %s", __func__, limit, turboRatioLimitConfigPath);
+                LOG("Turbo ratio limit is not a valid hexadecimal constant at %s", limit, turboRatioLimitConfigPath);
             } else {
                 uint64_t curLimit = rdmsr64(MSR_TURBO_RATIO_LIMIT);
                 uint64_t usrLimit = static_cast<uint64_t>(limit);
                 if (setIfNotEqual(curLimit, usrLimit, MSR_TURBO_RATIO_LIMIT)) {
-                    myLOG("%s: Change turbo ratio limit: 0x%llx -> 0x%llx", __func__, curLimit, usrLimit);
+                    LOG("Change turbo ratio limit: 0x%llx -> 0x%llx", curLimit, usrLimit);
                 }
             }
         }
     }
 
     if (ProcHotPath) {
-        bool prev = rdmsr64(MSR_IA32_POWER_CTL) & kEnableProcHotBit;
-        uint8_t *buffer = readFileNBytes(ProcHotPath, 0, 1);
-        // check if currently request enable ProcHot
-        bool curr = buffer && (*buffer == '1');
-        // deallocate the buffer
-        deleter(buffer);
-        if (curr != prev) {
-            myLOG("readConfigAtRunTime: %s Intel Proc Hot", curr ? "enable" : "disable");
-            if (curr) {
+        if (uint8_t *buffer = readFileAsBytes(ProcHotPath, 0, 1)) {
+            if (*buffer == '1') {
                 enableProcHot();
-            } else if((rdmsr64(MSR_IA32_MISC_ENABLE) & kEnableTurboBoostBits) != kEnableTurboBoostBits){
-                disableProcHot();
             } else {
-                myLOG("readConfigAtRuntime: Cannot disable PROCHOT while turboboost is active!");
+                disableProcHot();
             }
-                
+            deleter(buffer);
         }
     }
     
     // set hwp request value if hwp is enable
     if (cpu_info.supportedHWP && hwpRequestConfigPath) {
-        if (uint8_t *hex = readFileNBytes(hwpRequestConfigPath, 0, 10)) {
+        if (uint8_t *hex = readFileAsBytes(hwpRequestConfigPath, 0, 10)) {
             // hex is not NULL means the hwp request config exist
             // let's check if the hex is valid before writing to MSR
             long req = hexToInt(reinterpret_cast<char*>(hex));
             deleter(hex);
             if (req == ERANGE) {
-                myLOG("%s: HWP Request %s is not a valid hexadecimal constant at %s", __func__, hex, hwpRequestConfigPath);
+                LOG("HWP Request %s is not a valid hexadecimal constant at %s", hex, hwpRequestConfigPath);
             } else {
                 uint64_t curHWPRequest = rdmsr64(MSR_IA32_HWP_REQUEST);
                 uint64_t usrHWPRequest = static_cast<uint64_t>(req);
-               if (setIfNotEqual(curHWPRequest, usrHWPRequest, MSR_IA32_HWP_REQUEST)) {
-                   myLOG("%s: change MSR_IA32_HWP_REQUEST(0x%llx): 0x%llx -> 0x%llx", __func__, MSR_IA32_HWP_REQUEST, curHWPRequest, usrHWPRequest);
-               }
+                if (setIfNotEqual(curHWPRequest, usrHWPRequest, MSR_IA32_HWP_REQUEST)) {
+                    LOG("change MSR_IA32_HWP_REQUEST(0x%llx): 0x%llx -> 0x%llx", MSR_IA32_HWP_REQUEST, curHWPRequest, usrHWPRequest);
+                }
             }
         }
     }
     
     if (!hwpEnableOnceSet && cpu_info.supportedHWP && speedShiftPath) {
-        // check if previous speed shift is enabled
-        bool prev = rdmsr64(MSR_IA32_PM_ENABLE) == kEnableSpeedShiftBit;
-        uint8_t *buffer = readFileNBytes(speedShiftPath, 0, 1);
-        // check if currently request enable speed shift
-        bool curr = buffer && (*buffer == '1');
-        if (buffer && curr != prev) {
-            myLOG("readConfigAtRuntime: %s Intel Speed Shift", curr ? "enable" : "disable");
-            if (curr) {
+        if (uint8_t *buffer = readFileAsBytes(speedShiftPath, 0, 1)) {
+            if (*buffer == '1') {
                 enableSpeedShift();
             } else {
                 disableSpeedShift();
             }
+            deleter(buffer);
         }
-        // deallocate the buffer
-        deleter(buffer);
     }
-    sender->setTimeoutMS(updateInterval);
+    
+    // restart the timer
+    if (timerSource && !this->isInactive()) {
+        // Don't use sender here which will cause MBP8,2(SANDYBRIDGE) KP
+        timerSource->setTimeout(updateInterval);
+    }
 }
 
 bool CPUTune::setIfNotEqual(const uint64_t current, const uint64_t expect, const uint32_t msr) const {
@@ -273,9 +253,7 @@ void CPUTune::enableTurboBoost()
     // flip bit 38 to 0
     const uint64_t val = cur & kEnableTurboBoostBits;
     if (setIfNotEqual(cur, val, MSR_IA32_MISC_ENABLE)) {
-        myLOG("%s: change 0x%llx to 0x%llx in MSR_IA32_MISC_ENABLE(0x%llx)", __func__, cur, val, MSR_IA32_MISC_ENABLE);
-    } else {
-        myLOG("%s: 0x%llx in MSR_IA32_MISC_ENABLE(0x%llx) remains the same", __func__, cur, MSR_IA32_MISC_ENABLE);
+        LOG("change 0x%llx to 0x%llx in MSR_IA32_MISC_ENABLE(0x%llx)", cur, val, MSR_IA32_MISC_ENABLE);
     }
 }
 
@@ -285,9 +263,7 @@ void CPUTune::disableTurboBoost()
     // flip bit 38 to 1
     const uint64_t val = cur | kDisableTurboBoostBits;
     if (setIfNotEqual(cur, val, MSR_IA32_MISC_ENABLE)) {
-        myLOG("%s: change 0x%llx to 0x%llx in MSR_IA32_MISC_ENABLE(0x%llx)", __func__, cur, val, MSR_IA32_MISC_ENABLE);
-    } else {
-        myLOG("%s: 0x%llx in MSR_IA32_MISC_ENABLE(0x%llx) remains the same", __func__, cur, MSR_IA32_MISC_ENABLE);
+        LOG("change 0x%llx to 0x%llx in MSR_IA32_MISC_ENABLE(0x%llx)", cur, val, MSR_IA32_MISC_ENABLE);
     }
 }
 
@@ -296,9 +272,7 @@ void CPUTune::disableProcHot()
     const uint64_t cur = rdmsr64(MSR_IA32_POWER_CTL);
     const uint64_t val = cur & kDisableProcHotBit;
     if (setIfNotEqual(cur, val, MSR_IA32_POWER_CTL)) {
-        myLOG("%s: change 0x%llx to 0x%llx in MSR_IA32_POWERCTL(0x%llx)", __func__,cur, val, MSR_IA32_POWER_CTL);
-    } else {
-        myLOG("%s: 0x%llx in MSR_IA32_POWER_CTL(0x%llx) remains the same", __func__, cur, MSR_IA32_POWER_CTL);
+        LOG("change 0x%llx to 0x%llx in MSR_IA32_POWERCTL(0x%llx)", cur, val, MSR_IA32_POWER_CTL);
     }
 }
 
@@ -307,9 +281,7 @@ void CPUTune::enableProcHot()
     const uint64_t cur = rdmsr64(MSR_IA32_POWER_CTL);
     const uint64_t val = cur | kEnableProcHotBit;
     if(setIfNotEqual(cur, val, MSR_IA32_POWER_CTL)) {
-        myLOG("%s: change 0x%llx to 0x%llx in MSR_IA32_POWERCTL(0x%llx)", __func__, cur, val, MSR_IA32_POWER_CTL);
-    } else {
-        myLOG("%s: 0x%llx in MSR_IA32_POWER_CTL(0x%llx) remains the same", __func__, cur, MSR_IA32_POWER_CTL);
+        LOG("change 0x%llx to 0x%llx in MSR_IA32_POWERCTL(0x%llx)", cur, val, MSR_IA32_POWER_CTL);
     }
 }
 
@@ -317,9 +289,7 @@ void CPUTune::enableSpeedShift()
 {
     const uint64_t cur = rdmsr64(MSR_IA32_PM_ENABLE);
     if (setIfNotEqual(cur, kEnableSpeedShiftBit, MSR_IA32_PM_ENABLE)) {
-        myLOG("%s: change 0x%llx to 0x%llx in MSR_IA32_PM_ENABLE(0x%llx)", __func__, cur, kEnableSpeedShiftBit, MSR_IA32_PM_ENABLE);
-    } else {
-        myLOG("%s: 0x%llx in MSR_IA32_PM_ENABLE(0x%llx) remains the same", __func__, cur, MSR_IA32_PM_ENABLE);
+        LOG("change 0x%llx to 0x%llx in MSR_IA32_PM_ENABLE(0x%llx)", cur, kEnableSpeedShiftBit, MSR_IA32_PM_ENABLE);
     }
 }
 
@@ -327,9 +297,7 @@ void CPUTune::disableSpeedShift()
 {
     const uint64_t cur = rdmsr64(MSR_IA32_PM_ENABLE);
     if (setIfNotEqual(cur, kDisableSpeedShiftBit, MSR_IA32_PM_ENABLE)) {
-        myLOG("%s: change 0x%llx to 0x%llx in MSR_IA32_PM_ENABLE(0x%llx)", __func__, cur, kEnableSpeedShiftBit, MSR_IA32_PM_ENABLE);
-    } else {
-        myLOG("%s: 0x%llx in MSR_IA32_PM_ENABLE(0x%llx) remains the same", __func__, cur, MSR_IA32_PM_ENABLE);
+        LOG("change 0x%llx to 0x%llx in MSR_IA32_PM_ENABLE(0x%llx)", cur, kEnableSpeedShiftBit, MSR_IA32_PM_ENABLE);
     }
 }
 
@@ -346,24 +314,24 @@ void CPUTune::stop(IOService *provider)
     // restore the previous MSR_IA32 state
     const uint64_t cur_ctk = rdmsr64(MSR_IA32_POWER_CTL);
     if (setIfNotEqual(cur_ctk, org_MSR_IA32_POWER_CTL, MSR_IA32_POWER_CTL)) {
-        myLOG("stop: restore MSR_IA32_POWER_CTK from 0x%llx to 0x%llx", cur_ctk, org_MSR_IA32_POWER_CTL);
+        LOG("restore MSR_IA32_POWER_CTK from 0x%llx to 0x%llx", cur_ctk, org_MSR_IA32_POWER_CTL);
     }
     const uint64_t cur_misc = rdmsr64(MSR_IA32_MISC_ENABLE);
     if (setIfNotEqual(cur_misc, org_MSR_IA32_MISC_ENABLE, MSR_IA32_MISC_ENABLE)) {
-        myLOG("stop: restore MSR_IA32_MISC_ENABLE from 0x%llx to 0x%llx", cur_misc, org_MSR_IA32_MISC_ENABLE);
+        LOG("restore MSR_IA32_MISC_ENABLE from 0x%llx to 0x%llx", cur_misc, org_MSR_IA32_MISC_ENABLE);
     }
     const uint64_t cur_perf_ctl = rdmsr64(MSR_IA32_PERF_CTL);
     if (setIfNotEqual(cur_perf_ctl, org_MSR_IA32_PERF_CTL, MSR_IA32_PERF_CTL)) {
-        myLOG("stop: restore MSR_IA32_PERF_CTL from 0x%llx to 0x%llx", cur_perf_ctl, org_MSR_IA32_PERF_CTL);
+        LOG("restore MSR_IA32_PERF_CTL from 0x%llx to 0x%llx", cur_perf_ctl, org_MSR_IA32_PERF_CTL);
     }
     if (cpu_info.supportedHWP) {
         const uint64_t cur_pm_enable = rdmsr64(MSR_IA32_PM_ENABLE);
         if (setIfNotEqual(cur_pm_enable, org_MSR_IA32_PM_ENABLE, MSR_IA32_PM_ENABLE)) {
-            myLOG("stop: restore MSR_IA32_PM_ENABLE from 0x%llx to 0x%llx", cur_pm_enable, org_MSR_IA32_PM_ENABLE);
+            LOG("restore MSR_IA32_PM_ENABLE from 0x%llx to 0x%llx", cur_pm_enable, org_MSR_IA32_PM_ENABLE);
         }
         const uint64_t cur_hwp_req = rdmsr64(MSR_IA32_HWP_REQUEST);
         if (setIfNotEqual(cur_hwp_req, org_HWPRequest, MSR_IA32_HWP_REQUEST)) {
-            myLOG("%s: restore MSR_IA32_HWP_REQUEST(0x%llx) from 0x%llx to 0x%llx", __func__, cur_hwp_req, org_HWPRequest);
+            LOG("restore MSR_IA32_HWP_REQUEST(0x%llx) from 0x%llx to 0x%llx", cur_hwp_req, org_HWPRequest);
         }
     }
     super::stop(provider);