+++ CPUTuneCore/CPUTune/CPUTune.hpp
@@ -49,7 +49,7 @@ private:
     static constexpr uint64_t kDisableProcHotBit = 0xFFFFFFFE;
     static constexpr uint64_t kEnableProcHotBit = 0x1;
     
-    
+
     IOWorkLoop *myWorkLoop;
     IOTimerEventSource *timerSource;
     void readConfigAtRuntime(OSObject *owner, IOTimerEventSource *sender);